backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
pypdf
python-dotenv
openai
langchain-community
